Do not return deployment name from creation function
This information is already available from a parameter passed into the function;
there is no need to duplicate it or confuse which is the source of truth.

Whilst it may be possible for a mutating webhook to change this, such a webhook
would be very bad for human or kubectl-based interaction and should be challenged.